Added confirmation and cancellation sound effects to the start menu
Made it so that pressing the confirm button on one of the start menu
buttons plays the confirmation sound effect.

Made it so that presing the cancel button on one of the start menu
buttons highlights the Exit button and plays the cancallation sound
effect.

Added ButtonCount property to ButtonNavigatorBase and implemented it in
MainMenuButtonNavigator.

Added the prototype for the options menu.