Added functionality to the Options button
When the Options button is clicked, the options menu will pop up.

Added checks to each of the input system action delegates to prevent
them from being invoked when the GameObject is not active.

Added CoroutineRunner so that some coroutines can still be run when the
GameObject they are being started from is not active.

Added TranslateSmoothly.