Revamped the implementation of button navigators
A button navigator used to have to inherit from ButtonNavigatorBase and
implement the logic itself. This will lead to a lot of code duplication
if more types of menus were added because most of the logic would be the
same. To fix this, the ButtonNavigator component was added. It serves as
a general solution for keeping track of the currently selected ThButton.
Future button navigators and selectors will just have to reference this
component and perform rendering and stuff, without having to worry about
which button is currently selected. ButtonNavigatorBase and ThButton was
also modified to support vertical, horizontal and gridlike button
layouts by giving each ThButton a XY position.